# CBE: Don't emit unused undefined array literals (optimization)