The AIR for handling undefined arrays looks like:

```
  %9 = constant([20]u8, undefined)
  ...
  %7 = alloc(*[20]u8)
  %11!= store(%7, %9!)
 ```

Before this change, the CBE would emit:

```
 // Result of resolveInst on the RHS of the store
 static zig_u8 const t1[zig_as_u64(20)] = "\252\252\252\252\252\252\252\252\252\252\252\252\252\252\252\252\252\252\252\252";
 ...
 zig_u8 t0[zig_as_u64(20)];
 ...
 memset(&t0, zig_as_u8(0xaa), sizeof(zig_u8 [zig_as_u64(20)]));
```

`t1` is never actually used here, since `airStore` knows to render the store as a memset of undefined.

The array constants are now only resolved when they are not undefined, elminating all these string literals from the output.

This reduces `zig2.c` from 126.86MB to 116.71MB. `-fproc-stat-report` reports:

| File         | Total time | User mode time | Peak memory usage |
|--------------|------------|----------------|-------------------|
| zig2_old.obj | 62.79s     | 61.35s         | 2484MB            |
| zig2_new.obj | 58.06s     | 56.71s         | 2413MB            |